Invoke detect_duplicate_commits.py with python3 instead of python
The detect_duplicate_commits.py script in the bioc_git_transition repo was
migrated from Python 2 to Python 3 so needs to be invoked with python3
instead of python.

Also use it to check the last 30 commits (like the duplicate commits hook
does on the git.bioconductor.org server), instead of 10 commits only.